# docs/pushpress-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PushPress Platform API",
    "description": "The PushPress Platform API provides a comprehensive suite of endpoints designed \nto facilitate seamless integration with the PushPress ecosystem. \nThis API allows developers to interact programmatically with the PushPress platform, \nenabling the automation of various tasks and the extension of platform capabilities.\n",
    "version": "3.0.0"
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "API-KEY",
        "in": "header",
        "description": "Primary authentication mechanism for the platform"
      },
      "bearer": {
        "scheme": "bearer",
        "type": "http",
        "bearerFormat": "JWT",
        "description": "JWT Bearer Token primarily used to provision access to the platform"
      }
    },
...
  },
  "paths": {
...
    "/messages/email/send": {
      "post": {
        "operationId": "sendEmail",
        "summary": "Send an Email",
        "tags": [
          "Messages"
        ],
        "description": "Send an email from the PushPress platform. Note that only first party apps can send emails",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customer": {
                    "type": "string",
                    "description": "Uuid of the customer recipient"
                  },
                  "subject": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string",
                    "description": "The text email body"
                  },
                  "html": {
                    "type": "string",
                    "description": "The HTML email body"
                  },
                  "from": {
                    "type": "string",
                    "description": "The email sender name"
                  },
                  "replyTo": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "email"
                      },
                      {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "format": "email"
                        }
                      }
                    ],
                    "description": "One or more reply-to addresses"
                  },
                  "type": {
                    "type": "string",
                    "description": "Optional email type "
                  }
                },
                "required": [
                  "customer",
                  "subject",
                  "text",
                  "html",
                  "from"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "company-id",
            "description": "When using multitenant API keys, specify the company"
          }
        ],
        "x-speakeasy-group": "messages.email",
        "x-speakeasy-name-override": "send",
        "x-internal": true,
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "success"
                      ]
                    }
                  },
                  "required": [
                    "status"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Unauthorized"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "Forbidden"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.pushpress.com/v3",
      "description": "production",
      "x-speakeasy-server-id": "production"
    },
    {
      "url": "https://api.pushpressstage.com/v3",
      "description": "staging",
      "x-speakeasy-server-id": "staging"
    },
    {
      "url": "https://api.pushpressdev.com/v3",
      "description": "development",
      "x-speakeasy-server-id": "development"
    }
  ],
...
  "security": [
    {
      "apiKey": []
    }
  ],
...
}